jumper/bot/requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
ccxt==4.2.53
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
Naked==0.1.32
pycares==4.4.0
pycparser==2.21
pycryptodome==3.20.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
